requirements.txt
gradio
keras>=3.0.0
keras-hub
jax
matplotlib
pillow
numpy
